requirements.txt
streamlit
pandas
pytesseract
Pillow
plotly
openpyxl
